requirements.txt
flask
PyGithub
github_webhook
